LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-getComputedStyle-002-expected.txt

Lorem ipsum dolor sit amet.
Lorem ipsum dolor sit amet.

FAIL getComputedStyle() with fragmented containing block in multicolumn layout assert_equals: expected "-85px" but got "25px"
PASS getComputedStyle() with fragmented containing block in inline layout
